refactor: id 타입 변경

+++ app/src/main/java/woowacourse/shopping/domain/repository/CartRepository.kt
@@ -8,5 +8,5 @@ interface CartRepository {
 
     suspend fun getAllCartProducts(): List<CartProduct>
 
-    suspend fun deleteCartProduct(id: Int)
+    suspend fun deleteCartProduct(id: Long)
 }
